@@ -29,68 +29,68 @@
     "seeder:run": "mikro-orm seeder:run"
   },
   "dependencies": {
-    "@mikro-orm/cli": "^5.7.13",
-    "@mikro-orm/core": "^5.7.13",
-    "@mikro-orm/migrations": "^5.7.13",
-    "@mikro-orm/nestjs": "^5.2.0",
-    "@mikro-orm/postgresql": "^5.7.13",
-    "@mikro-orm/seeder": "^5.7.13",
+    "@mikro-orm/cli": "^5.8.3",
+    "@mikro-orm/core": "^5.8.3",
+    "@mikro-orm/migrations": "^5.8.3",
+    "@mikro-orm/nestjs": "^5.2.2",
+    "@mikro-orm/postgresql": "^5.8.3",
+    "@mikro-orm/seeder": "^5.8.3",
     "@mikro-orm/sql-highlighter": "^1.0.1",
-    "@nestjs/common": "^10.0.5",
-    "@nestjs/config": "^3.0.0",
-    "@nestjs/core": "^10.0.5",
-    "@nestjs/jwt": "^10.1.0",
-    "@nestjs/passport": "^10.0.0",
-    "@nestjs/platform-express": "^10.0.5",
-    "@nestjs/swagger": "^7.1.1",
-    "bcrypt": "^5.1.0",
+    "@nestjs/common": "^10.2.6",
+    "@nestjs/config": "^3.1.1",
+    "@nestjs/core": "^10.2.6",
+    "@nestjs/jwt": "^10.1.1",
+    "@nestjs/passport": "^10.0.2",
+    "@nestjs/platform-express": "^10.2.6",
+    "@nestjs/swagger": "^7.1.12",
+    "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
-    "country-state-city": "^3.1.4",
-    "flat": "^5.0.2",
-    "joi": "^17.9.2",
+    "country-state-city": "^3.2.1",
+    "flat": "^6.0.1",
+    "joi": "^17.10.2",
     "lodash": "^4.17.21",
     "moment": "^2.29.4",
-    "nestjs-i18n": "^10.2.6",
+    "nestjs-i18n": "^10.3.5",
     "passport": "^0.6.0",
     "passport-jwt": "^4.0.1",
     "reflect-metadata": "^0.1.13",
     "rxjs": "^7.8.1",
-    "uuid": "^9.0.0"
+    "uuid": "^9.0.1"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.6.6",
-    "@commitlint/config-conventional": "^17.6.6",
-    "@nestjs/cli": "^10.1.8",
-    "@nestjs/schematics": "^10.0.1",
-    "@nestjs/testing": "^10.0.5",
+    "@commitlint/cli": "^17.7.1",
+    "@commitlint/config-conventional": "^17.7.0",
+    "@nestjs/cli": "^10.1.18",
+    "@nestjs/schematics": "^10.0.2",
+    "@nestjs/testing": "^10.2.6",
     "@types/bcrypt": "^5.0.0",
-    "@types/express": "^4.17.17",
-    "@types/flat": "^5.0.2",
-    "@types/jest": "29.5.3",
-    "@types/lodash": "^4.14.195",
+    "@types/express": "^4.17.18",
+    "@types/flat": "^5.0.3",
+    "@types/jest": "29.5.5",
+    "@types/lodash": "^4.14.199",
     "@types/multer": "^1.4.7",
-    "@types/node": "20.4.2",
-    "@types/passport-jwt": "^3.0.9",
-    "@types/supertest": "^2.0.12",
-    "@types/uuid": "^9.0.2",
-    "@typescript-eslint/eslint-plugin": "^6.0.0",
-    "@typescript-eslint/parser": "^6.0.0",
-    "eslint": "^8.45.0",
-    "eslint-config-prettier": "^8.8.0",
+    "@types/node": "20.6.5",
+    "@types/passport-jwt": "^3.0.10",
+    "@types/supertest": "^2.0.13",
+    "@types/uuid": "^9.0.4",
+    "@typescript-eslint/eslint-plugin": "^6.7.2",
+    "@typescript-eslint/parser": "^6.7.2",
+    "eslint": "^8.50.0",
+    "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-prettier": "^5.0.0",
     "eslint-plugin-simple-import-sort": "^10.0.0",
     "husky": "^8.0.3",
-    "jest": "29.6.1",
-    "lint-staged": "^13.2.3",
-    "prettier": "^3.0.0",
+    "jest": "29.7.0",
+    "lint-staged": "^14.0.1",
+    "prettier": "^3.0.3",
     "source-map-support": "^0.5.21",
     "supertest": "^6.3.3",
     "ts-jest": "29.1.1",
     "ts-loader": "^9.4.4",
     "ts-node": "^10.9.1",
     "tsconfig-paths": "4.2.0",
-    "typescript": "^5.1.6"
+    "typescript": "^5.2.2"
   },
   "mikro-orm": {
     "useTsNode": true,
